refactor: SearchBar
The search bar has been moved. Now it is in the components folder
