+++ lib/pages/mapPage.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
+import 'package:postin_app/components/search_bar/searchBar.dart';
 
 class MapPage extends StatelessWidget {
   @override
@@ -46,50 +47,4 @@ class MapPage extends StatelessWidget {
         urlTemplate: 'https://tile.openstreetmap.org/{z}/{x}/{y}.png',
         userAgentPackageName: 'dev.fleaflet.flutter_map.exemple',
       );
-
-  Widget buildSearchField(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.fromLTRB(
-          16, 35, 16, 15), // Ajusta el padding superior aquí
-      child: Theme(
-        data: Theme.of(context).copyWith(
-          inputDecorationTheme: InputDecorationTheme(
-            filled: true,
-            fillColor: Theme.of(context).primaryColor.withOpacity(0.1),
-            border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(40),
-              borderSide:
-                  BorderSide(color: Theme.of(context).primaryColor, width: 2),
-            ),
-            focusedBorder: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(40),
-              borderSide:
-                  BorderSide(color: Theme.of(context).primaryColor, width: 2),
-            ),
-            labelStyle: TextStyle(
-                color: Theme.of(context)
-                    .primaryColor), // Cambia el color de la etiqueta
-            hintStyle: TextStyle(
-                color: Theme.of(context).primaryColor.withOpacity(
-                    0.5)), // Cambia el color del texto de sugerencia
-          ),
-          textSelectionTheme: TextSelectionThemeData(
-            cursorColor:
-                Theme.of(context).primaryColor, // Cambia el color del cursor
-          ),
-        ),
-        child: TextFormField(
-          autofocus: false,
-          autofillHints: [AutofillHints.email],
-          obscureText: false,
-          decoration: InputDecoration(
-            labelText: 'Post-In',
-            contentPadding: EdgeInsets.symmetric(vertical: 15, horizontal: 24),
-            suffixIcon: Icon(Icons.filter_list_outlined),
-          ),
-          keyboardType: TextInputType.emailAddress,
-        ),
-      ),
-    );
-  }
 }
